Improve-flask-template
