+++ src/universalinit/templates/flask/app/__init__.py
@@ -1,6 +1,6 @@
 from flask import Flask
-from flask_smorest import Api, Blueprint
-from flask.views import MethodView
+from .routes.health import blp
+from flask_smorest import Api
 
 
 app = Flask(__name__)
@@ -8,16 +8,8 @@ app.config["API_TITLE"] = "My Flask API"
 app.config["API_VERSION"] = "v1"
 app.config["OPENAPI_VERSION"] = "3.0.3"
 app.config['OPENAPI_URL_PREFIX'] = '/docs'
-app.config["OPENAPI_SWAGGER_UI_PATH"] = "/"
+app.config["OPENAPI_SWAGGER_UI_PATH"] = ""
 app.config["OPENAPI_SWAGGER_UI_URL"] = "https://cdn.jsdelivr.net/npm/swagger-ui-dist/"
-
-blp = Blueprint("Healt Check", "health check", url_prefix="/", description="Health check route")
-
-
-@blp.route("/")
-class HealthCheck(MethodView):
-    def get(self):
-        return {"message": "Healthy"}
 
 
 api = Api(app)
